# sgemm: Reduce unnecessary AVX register permutations