- Removed redundant `_mm256_permute2f128_ps` instructions for lane swapping.
- Reordered final output assignments to match the expected layout directly, simplifying downstream processing.
- This change reduces register pressure and improves instruction efficiency without altering the computation logic.

### Performance Benchmarks:
**Tested on Intel 10700k (Single core)**  
| Matrix Size | Before Optimization (ns/iter) | After Optimization (ns/iter) | Improvement (%) |
|--------------|-------------------------------|------------------------------|-----------------|
| m004        | 177 (+/- 1)                  | 174 (+/- 17)                | **1.69%**      |
| m006        | 209 (+/- 19)                 | 202 (+/- 20)                | **3.35%**      |
| m008        | 200 (+/- 0)                  | 191 (+/- 1)                 | **4.50%**      |
| m012        | 267 (+/- 2)                  | 243 (+/- 4)                 | **8.99%**      |
| m016        | 278 (+/- 6)                  | 233 (+/- 10)                | **16.19%**     |
| m032        | 1,344 (+/- 14)               | 989 (+/- 41)                | **26.39%**     |
| m064        | 8,974 (+/- 169)              | 6,050 (+/- 17)              | **32.60%**     |
| m127        | 64,466 (+/- 1,468)           | 41,704 (+/- 78)             | **35.32%**     |