@bluss 2025-05-11
If you are curious, larger matrix benchmarks are available in the other benchmark script.

Like (`xsv` just to make a table, optional)

```
./benches/benchloop.py  -t f32 -s 384 450 512 | xsv table
```