@SongXiaoXi 2025-05-11
I ran your script and obtained the following results:  
Before:
| m   | k   | n   | layout | type | average\_ns | minimum\_ns | median\_ns | samples | GFLOPS | nc | kc | mc | threads |
| --- | --- | --- | ------ | ---- | ----------- | ----------- | ---------- | ------- | ------ | -- | -- | -- | ------- |
| 384 | 384 | 384 | FCC    | f32  | 1,618,910   | 1,617,871   | 1,618,362  | 1560    | 69.95213322544181  |    |    |    | 0       |
| 450 | 450 | 450 | FCC    | f32  | 2,650,805   | 2,647,933   | 2,648,578  | 1560    | 68.75269965161526  |    |    |    | 0       |
| 512 | 512 | 512 | FCC    | f32  | 3,834,974   | 3,821,840   | 3,824,136  | 310     | 69.99668211570665  |    |    |    | 0       |


After:
| m   | k   | n   | layout | type | average\_ns | minimum\_ns | median\_ns | samples | GFLOPS     | nc | kc | mc | threads |
| --- | --- | --- | ------ | ---- | ----------- | ----------- | ---------- | ------- | ---------- | -- | -- | -- | ------- |
| 384 | 384 | 384 | FCC    | f32  | 974,601     | 973,492     | 974,549    | 1560    | 116.1975085188708 |    |    |    | 0       |
| 450 | 450 | 450 | FCC    | f32  | 1,594,266   | 1,584,060   | 1,590,237  | 1560    | 114.31592971310936 |    |    |    | 0       |
| 512 | 512 | 512 | FCC    | f32  | 2,334,892   | 2,289,947   | 2,357,830  | 1560    | 114.96696892190303 |    |    |    | 0       |