@bluss 2025-05-11
thanks for this huge improvement. Make sure to ping me if I don't get back to releasing this in the next week